requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
Django==5.2.5
fastapi==0.119.0
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
